requirements.txt
amqp==5.1.1
anyio==3.5.0
asgiref==3.3.1
attrs==21.4.0
billiard==3.6.4.0
cachetools==4.2.1
celery==5.2.6
certifi==2020.12.5
chardet==3.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.4
django-crispy-forms==1.10.0
django-datatables-too==0.1.3
django-environ==0.4.5
django-extra-views==0.13.0
django-filter==2.4.0
django-multi-form-view==2.0.1
django-rest-authtoken==2.1.4
django-rest-swagger==2.2.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
fastapi==0.76.0
google-api-core==1.26.3
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.5.1
kombu==5.2.4
MarkupSafe==2.1.1
numpy==1.22.3
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.9
Pillow==9.1.0
prompt-toolkit==3.0.29
protobuf==3.15.6
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.1
pytz==2022.1
PyYAML==6.0
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.1
starlette==0.18.0
stripe==2.55.1
swagger-spec-validator==2.7.4
typing_extensions==4.2.0
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
